packages/akshare-mcp/requirements.txt
mcp>=1.0.0
akshare>=1.10.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
numba>=0.59.0
asyncpg>=0.29.0
pandas-ta>=0.3.14
TA-Lib>=0.4.28
tushare>=1.4.0
baostock>=0.8.8
efinance>=0.5.5
pydantic>=2.0.0
ray>=2.9.0
pytest-benchmark>=4.0.0
